refactor: Admin 폴더 정리
11월 27일 회의 내용에 따라 폴더를 다음과 같이 정리하였습니다.
- 페이지 역할을 하는 컴포넌트만 페이지 폴더에 남겨놓는다.
컴포넌트 폴더로 이동한 컴포넌트들
- MenuStatusBar
- StoreCreateStatusBar
- StoreUpdateStatusBar
- AdminStoreListHeader
- ProductListItem
- StoreListItem
- AdminLoadingAndGetUser

페이지에는 페이지 역할을 하는 컴포넌트만 남겨져있으며 styles를 하위 도메인별로 통합하였습니다.
