+++ src/Components/Bar/styles.ts
@@ -1,0 +1,51 @@
+import { SubTitle1 } from "@/lib/styles";
+import styled from "styled-components";
+
+export const StatusBar = styled.div`
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  padding: 0 1rem;
+  height: 8rem;
+
+  .status-bar-item-container {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    height: 100%;
+    border-top: 1px solid ${(props) => props.theme.color.gray300};
+  }
+  .status-message-container {
+    h3 {
+      ${SubTitle1};
+      color: ${(props) => props.theme.color.fontColorBlack};
+    }
+  }
+  .status-button-container {
+    button {
+      cursor: pointer;
+      padding: 0.5rem 1.8rem;
+      border: 0;
+      font-size: 2rem;
+      color: ${(props) => props.theme.color.fontColorWhite};
+      border-radius: 0.2rem;
+      line-height: 2.8rem;
+    }
+
+    .cancel-button {
+      background-color: ${(props) => props.theme.color.gray300};
+    }
+    .confirm-button {
+      margin-left: 0.5rem;
+      background-color: ${(props) => props.theme.color.primary700};
+    }
+  }
+  ${({ theme }) => theme.device.tablet} {
+    height: 10rem;
+    .status-bar-item-container {
+      flex-direction: column;
+      justify-content: space-around;
+    }
+  }
+`;
