+++ src/Components/Layouts/Header/AdminStoreListHeader.tsx
@@ -1,29 +1,29 @@
 import { PageHeader } from "@/Components";
-import { storeStateProps, userState } from "@/lib/state";
+import { storeStateProps } from "@/lib/state";
 import { MdAddCircle } from "react-icons/md";
-import { Link } from "react-router-dom";
-import { useRecoilValue } from "recoil";
-import { AddStoreButton, Header } from "../styles";
+import { Link, useParams } from "react-router-dom";
+
+import { AddStoreButton, Header } from "../../../Page/Admin/Store/styles";
 
 interface Props {
-  store: storeStateProps[];
+  store?: storeStateProps[];
 }
 
 const AdminStoreListHeader = ({ store }: Props) => {
-  const user = useRecoilValue(userState);
+  const { userId } = useParams();
   return (
     <Header>
       <PageHeader
         header="가게목록"
         message={
-          store.length !== 0
+          store?.length !== 0
             ? "가게를 선택하세요."
             : "아직 등록된 가게가 없습니다."
         }
       />
       <AddStoreButton>
         <MdAddCircle />
-        <Link to={`/admin/${user.id}/store/create`}>가게등록</Link>
+        <Link to={`/admin/${userId}/store/create`}>가게등록</Link>
       </AddStoreButton>
     </Header>
   );
