+++ src/Components/Layouts/Header/index.ts
@@ -1,1 +1,2 @@
 export { default as PageHeader } from "./PageHeader";
+export { default as AdminStoreListHeader } from "./AdminStoreListHeader";
