+++ src/Components/Layouts/index.ts
@@ -3,4 +3,4 @@ export { default as AdminManageProductLayout } from "./AdminManageProductLayout"
 export { default as ClientLayout } from "./ClientLayout";
 export { default as LandingLayout } from "./LandingLayout";
 export { default as OrderLayout } from "./OrderLayout";
-export { PageHeader } from "./Header";
+export { PageHeader, AdminStoreListHeader } from "./Header";
