+++ src/Components/List/ListItem/index.ts
@@ -1,0 +1,2 @@
+export { default as ProductListItem } from "./ProductListItem";
+export { default as StoreListItem } from "./StoreListItem";
