+++ src/Components/Loading/index.ts
@@ -1,2 +1,3 @@
 export { default as Loading } from "./Loading";
 export { default as LoadingBall } from "./LoadingBall";
+export { default as AdminLoadingAndGetUser } from "./AdminLoadingAndGetUser";
