+++ src/Page/Admin/Login/index.ts
@@ -1,2 +1,1 @@
-export { default as AdminLoadingAndGetUser } from "./AdminLoadingAndGetUser";
-export { default as AdminLogin } from "./AdminLogin";
+export { default as Login } from "./Login";
