--- src/Page/Admin/Product/AdminManageProductAddItem/index.ts
@@ -1,1 +1,0 @@
-export { default as AdminManageProductAddItem } from "./AdminManageProductAddItem";
