--- src/Page/Admin/Product/AdminManageProductItem/index.ts
@@ -1,1 +1,0 @@
-export { default as AdminManageProductItemList } from "./AdminManageProductItemList";
