--- src/Page/Admin/Product/AdminManageProductMain/index.ts
@@ -1,1 +1,0 @@
-export { default as AdminManageProductMain } from "./AdminManageProductMain";
