--- src/Page/Admin/Product/AdminProductDetail/index.ts
@@ -1,1 +1,0 @@
-export { default as AdminProductDetail } from "./AdminProductDetail";
