--- src/Page/Admin/Product/StateMenuBar/index.ts
@@ -1,1 +1,0 @@
-export { default as StateMenuBar } from "./StateMenuBar";
