+++ src/Page/Admin/Product/index.ts
@@ -1,0 +1,4 @@
+export { default as CreateProductPage } from "./CreateProductPage";
+export { default as ProductManageMainPage } from "./ProductManageMainPage";
+export { default as ProductListPage } from "./ProductListPage";
+export { default as ProductDetailPage } from "./ProductDetailPage";
