--- src/Page/Admin/Store/AdminStoreCreate/index.ts
@@ -1,1 +1,0 @@
-export { default as AdminStoreCreate } from "./AdminStoreCreate";
