--- src/Page/Admin/Store/AdminStoreList/index.ts
@@ -1,1 +1,0 @@
-export { default as AdminStoreList } from "./AdminStoreList";
