--- src/Page/Admin/Store/AdminStoreUpdate/index.ts
@@ -1,1 +1,0 @@
-export { default as AdminStoreUpdate } from "./AdminStoreUpdate";
