+++ src/Page/Admin/Store/styles.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { ButtonDefault } from "@/Components";
-import { Body1, Headline3, SubTitle1, SubTitle2 } from "@/lib/styles/mixin";
+import { Body1, SubTitle1, SubTitle2 } from "@/lib/styles/mixin";
 
 export const Wrapper = styled.div``;
 export const Header = styled.div`
@@ -48,49 +48,6 @@ export const StoreList = styled.ul`
   }
 `;
 
-export const Item = styled.li`
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  height: inherit;
-  position: relative;
-  border: 1px solid ${(props) => props.theme.color.gray300};
-  border-radius: 1rem;
-  padding: 0.5rem 1rem;
-  a {
-    text-decoration: none;
-    color: ${(props) => props.theme.color.fontColorBlack};
-  }
-  h3 {
-    ${Headline3};
-    padding: 0;
-    margin: 0;
-    ${({ theme }) => theme.device.mobile} {
-      font-size: 2.4rem;
-    }
-  }
-`;
-
-export const ButtonContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  .various-button-box {
-    button {
-      cursor: pointer;
-      border: 0;
-      background-color: unset;
-      padding: 0;
-      font-size: 2rem;
-      color: ${(props) => props.theme.color.gray400};
-    }
-  }
-`;
-export const UpdateButton = styled(ButtonDefault)``;
-export const DeleteButton = styled(ButtonDefault)`
-  margin-left: 1.8rem;
-`;
-
 export const Form = styled.form`
   width: 100%;
   padding: 1rem 0;
@@ -133,55 +90,6 @@ export const InputContainer = styled.div<{ disabled?: boolean }>`
       grid-column: 1 / 10;
       ${SubTitle2};
       color: ${(props) => props.theme.color.error500};
-    }
-  }
-`;
-
-export const StatusBar = styled.div`
-  position: fixed;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  padding: 0 1rem;
-  height: 8rem;
-
-  .status-bar-item-container {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 100%;
-    border-top: 1px solid ${(props) => props.theme.color.gray300};
-  }
-  .status-message-container {
-    h3 {
-      ${SubTitle1};
-      color: ${(props) => props.theme.color.fontColorBlack};
-    }
-  }
-  .status-button-container {
-    button {
-      cursor: pointer;
-      padding: 0.5rem 1.8rem;
-      border: 0;
-      font-size: 2rem;
-      color: ${(props) => props.theme.color.fontColorWhite};
-      border-radius: 0.2rem;
-      line-height: 2.8rem;
-    }
-
-    .cancel-button {
-      background-color: ${(props) => props.theme.color.gray300};
-    }
-    .confirm-button {
-      margin-left: 0.5rem;
-      background-color: ${(props) => props.theme.color.primary700};
-    }
-  }
-  ${({ theme }) => theme.device.tablet} {
-    height: 10rem;
-    .status-bar-item-container {
-      flex-direction: column;
-      justify-content: space-around;
     }
   }
 `;
