--- src/Page/Admin/index.ts
@@ -1,1 +1,0 @@
-export { default as Logout } from "./Logout/Logout";
