--- src/Page/index.ts
@@ -1,1 +1,0 @@
-export { Logout } from "./Admin";
